+++ src/types/type_name/mod.rs
@@ -6,7 +6,7 @@ mod set;
 #[allow(clippy::module_inception)]
 mod type_name;
 
+pub use crate::wrap::TypeName;
 pub use display_mode::{DisplayMode, Full, Short};
 pub use list::TypeNameList;
 pub use set::{TypeNameMap, TypeNameSet};
-pub use type_name::TypeName;