+++ src/wrap/mod.rs
@@ -5,9 +5,11 @@ mod display_as_debug;
 mod opaque;
 mod option;
 mod result;
+mod type_name;
 
 pub use debug_as_display::DebugAsDisplay;
 pub use display_as_debug::DisplayAsDebug;
 pub use opaque::Opaque;
 pub use option::{OpaqueOption, TypeNameOption};
 pub use result::{OpaqueResult, TypeNameResult};
+pub use type_name::TypeName;